Support running tests in embassy-stm32 and move impl from common back to stm32
