--- embassy-hal-common/src/stm32/flash/mod.rs
@@ -1,2 +1,0 @@
-pub mod f4;
-pub mod f7;
