--- embassy-hal-common/src/stm32/mod.rs
@@ -1,1 +1,0 @@
-pub mod flash;
